backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
drf-yasg==1.21.7
psycopg2-binary==2.9.0
channels==4.3.1
channels_radis==4.3.0
